Split componentMocking.tsx into smaller files
componentMocking.tsx grew too much and tried to do much more than name suggests. Instead of importing componentMocking.tsx, now use testingLibraryTools.ts

React components are moved their own files
ExposeHooksForTesting.tsx has hook test utils
renderComponentTestDOM creates the DOM with mockProviders and contexts.

No new code added, just splitted and removed componentMocking.tsx.
