+++ src/common/test/RenderChildrenWhenDataIsComplete.tsx
@@ -1,0 +1,17 @@
+import React, { useContext } from 'react';
+
+import { ProfileContext } from '../../profile/context/ProfileContext';
+
+function RenderChildrenWhenDataIsComplete({
+  children,
+}: {
+  children: React.ReactElement | React.ReactNodeArray;
+}): React.ReactElement | null {
+  const { isComplete } = useContext(ProfileContext);
+  if (!isComplete) {
+    return null;
+  }
+  return <div data-testid="component-wrapper">{children}</div>;
+}
+
+export default RenderChildrenWhenDataIsComplete;
